fix: cfb
